Fix bugs & RMF frame eval
Flipping (bi)normal is a proper result from my approach - when the 'reference up direction' slightly swings from one side to the other side of the curve tangent, their cross product, or the normal vector, flips.

The way to solve this is to use RMF or Optimal Transport. Gemini suggests RMF, so I accept it. The result looks good to me.
For the RMF algorithm, see "Computation of Rotation Minimizing Frames."